+++ utils/deviceInfo.js
